+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # All changes will be documented here.
+
+- Support basic configuration for an EC2 instance.